feat: improve lens display logic (#9558)
* feat: improve lens display logic

* chore: reply review

* chore: reply review
