+++ packages/shared/src/locales/en-US.json
@@ -191,6 +191,8 @@
     "applications_create_persona_hint": "Please create a Persona and verify your account to use this function.",
     "applications_create_persona_action": "Create persona",
     "lens_follow": "Follow",
+    "lens_following": "Following",
+    "view": "View",
     "verified_by": "Verified by {{ marketplace }}",
     "no_NFTs_found": "No NFTs found."
 }
